Prevent race condition when using taskrun and pipelinerun attestation
If the last task in the pipeline is the one that builds an image (thus provides the expected type-hinting), there's a race condition which causes either the taskrun or the pipelinerun attestation to be present in the registry. This change confirms the taskruns have been reconciled in chains before reconciling the pipelinerun.